Extract our usage of the Drawer to a local component for easier reuse
This will make it easier to have different menu items for customer,
steward, and signator.
