+++ src/components/index.ts
@@ -1,6 +1,7 @@
 import filteredList from './filteredList'
 import table from './table'
 import App from './App.svelte'
+import AppDrawer from './AppDrawer.svelte'
 import AppFooter from './AppFooter.svelte'
 import Banner from './Banner.svelte'
 import Breadcrumb from './Breadcrumb.svelte'
@@ -28,6 +29,7 @@ export {
   filteredList as FilteredList,
   table as Table,
   App,
+  AppDrawer,
   AppFooter,
   Banner,
   Breadcrumb,
