+++ src/pages/_layout.svelte
@@ -1,11 +1,6 @@
 <script lang="ts">
 import user, { isSteward } from '../authn/user'
-import AppFooter from '../components/AppFooter.svelte'
-import AppHeader from '../components/AppHeader.svelte'
-import { goto } from '@roxi/routify'
-import { Drawer } from '@silintl/ui-components'
-
-let toggle = false
+import { AppDrawer } from '../components/'
 
 $: menuItems = [
   {},
@@ -47,18 +42,8 @@ $: menuItems = [
     button: true,
   },
 ]
-
-const logoClickHandler = () => $goto('/')
 </script>
 
-<Drawer miniMenu modal hideForPhonesOnly {toggle} {menuItems} title="Covered" class="auto-width border-white">
-  <span class="pointer" on:click={logoClickHandler} slot="header">
-    <img class="w-100" src="/logo.svg" alt="Cover" />
-  </span>
-
-  <AppHeader on:toggleDrawer={() => (toggle = !toggle)} />
-
+<AppDrawer {menuItems}>
   <slot />
-
-  <AppFooter />
-</Drawer>
+</AppDrawer>
